fix: validate oauth redirects
Amp-Thread-ID: https://ampcode.com/threads/T-019e4f9d-104f-72b9-be8b-d3edd5205b01
